// server/src/generated/gtfs-realtime.ts
export interface FeedMessage {
    /**
     * Metadata about this feed and feed message.
     *
     * @generated from protobuf field: required transit_realtime.FeedHeader header = 1
     */
    header?: FeedHeader;
    /**
     * Contents of the feed.
     *
     * @generated from protobuf field: repeated transit_realtime.FeedEntity entity = 2
     */
    entity: FeedEntity[];
}
/**
 * Metadata about a feed, included in feed messages.
 *
 * @generated from protobuf message transit_realtime.FeedHeader
 */
export interface FeedHeader {
    /**
     * Version of the feed specification.
     * The current version is 2.0.  Valid versions are "2.0", "1.0".
     *
     * @generated from protobuf field: required string gtfs_realtime_version = 1
     */
    gtfsRealtimeVersion: string;
    /**
     * @generated from protobuf field: optional transit_realtime.FeedHeader.Incrementality incrementality = 2 [default = FULL_DATASET]
     */
    incrementality?: FeedHeader_Incrementality;
    /**
     * This timestamp identifies the moment when the content of this feed has been
     * created (in server time). In POSIX time (i.e., number of seconds since
     * January 1st 1970 00:00:00 UTC).
     *
     * @generated from protobuf field: optional uint64 timestamp = 3
     */
    timestamp?: bigint;
    /**
     * String that matches the feed_info.feed_version from the GTFS feed that the real
     * time data is based on. Consumers can use this to identify which GTFS feed is
     * currently active or when a new one is available to download.
     *
     * @generated from protobuf field: optional string feed_version = 4
     */
    feedVersion?: string;
}
// ...
export enum FeedHeader_Incrementality {
    /**
     * @generated from protobuf enum value: FULL_DATASET = 0;
     */
    FULL_DATASET = 0,
    /**
     * @generated from protobuf enum value: DIFFERENTIAL = 1;
     */
    DIFFERENTIAL = 1
}
/**
 * A definition (or update) of an entity in the transit feed.
 *
 * @generated from protobuf message transit_realtime.FeedEntity
 */
export interface FeedEntity {
    /**
     * The ids are used only to provide incrementality support. The id should be
     * unique within a FeedMessage. Consequent FeedMessages may contain
     * FeedEntities with the same id. In case of a DIFFERENTIAL update the new
     * FeedEntity with some id will replace the old FeedEntity with the same id
     * (or delete it - see is_deleted below).
     * The actual GTFS entities (e.g. stations, routes, trips) referenced by the
     * feed must be specified by explicit selectors (see EntitySelector below for
     * more info).
     *
     * @generated from protobuf field: required string id = 1
     */
    id: string;
    /**
     * Whether this entity is to be deleted. Relevant only for incremental
     * fetches.
     *
     * @generated from protobuf field: optional bool is_deleted = 2 [default = false]
     */
    isDeleted?: boolean;
    /**
     * Data about the entity itself. Exactly one of the following fields must be
     * present (unless the entity is being deleted).
     *
     * @generated from protobuf field: optional transit_realtime.TripUpdate trip_update = 3
     */
    tripUpdate?: TripUpdate;
    /**
     * @generated from protobuf field: optional transit_realtime.VehiclePosition vehicle = 4
     */
    vehicle?: VehiclePosition;
    /**
     * @generated from protobuf field: optional transit_realtime.Alert alert = 5
     */
    alert?: Alert;
    /**
     * NOTE: This field is still experimental, and subject to change. It may be formally adopted in the future.
     *
     * @generated from protobuf field: optional transit_realtime.Shape shape = 6
     */
    shape?: Shape;
    /**
     * @generated from protobuf field: optional transit_realtime.Stop stop = 7
     */
    stop?: Stop;
    /**
     * @generated from protobuf field: optional transit_realtime.TripModifications trip_modifications = 8
     */
    tripModifications?: TripModifications;
}
// ...
export interface Shape {
    /**
     * Identifier of the shape. Must be different than any shape_id defined in the (CSV) GTFS.
     * This field is required as per reference.md, but needs to be specified here optional because "Required is Forever"
     * See https://developers.google.com/protocol-buffers/docs/proto#specifying_field_rules
     * NOTE: This field is still experimental, and subject to change. It may be formally adopted in the future.
     *
     * @generated from protobuf field: optional string shape_id = 1
     */
    shapeId?: string;
    /**
     * Encoded polyline representation of the shape. This polyline must contain at least two points and represent the full shape of the trip where it's used.
     * For more information about encoded polylines, see https://developers.google.com/maps/documentation/utilities/polylinealgorithm
     * This field is required as per reference.md, but needs to be specified here optional because "Required is Forever"
     * See https://developers.google.com/protocol-buffers/docs/proto#specifying_field_rules
     * NOTE: This field is still experimental, and subject to change. It may be formally adopted in the future.
     *
     * @generated from protobuf field: optional string encoded_polyline = 2
     */
    encodedPolyline?: string;
}
// ...
class FeedMessage$Type extends MessageType<FeedMessage> {
    constructor() {
        super("transit_realtime.FeedMessage", [
            { no: 1, name: "header", kind: "message", T: () => FeedHeader },
            { no: 2, name: "entity", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => FeedEntity }
        ]);
    }
    create(value?: PartialMessage<FeedMessage>): FeedMessage {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.entity = [];
        if (value !== undefined)
            reflectionMergePartial<FeedMessage>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: FeedMessage): FeedMessage {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* required transit_realtime.FeedHeader header */ 1:
                    message.header = FeedHeader.internalBinaryRead(reader, reader.uint32(), options, message.header);
                    break;
                case /* repeated transit_realtime.FeedEntity entity */ 2:
                    message.entity.push(FeedEntity.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: FeedMessage, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* required transit_realtime.FeedHeader header = 1; */
        if (message.header)
            FeedHeader.internalBinaryWrite(message.header, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* repeated transit_realtime.FeedEntity entity = 2; */
        for (let i = 0; i < message.entity.length; i++)
            FeedEntity.internalBinaryWrite(message.entity[i], writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message transit_realtime.FeedMessage
 */
export const FeedMessage = new FeedMessage$Type();
// @generated message type with reflection information, may provide speed optimized methods
class FeedHeader$Type extends MessageType<FeedHeader> {
    constructor() {
        super("transit_realtime.FeedHeader", [
            { no: 1, name: "gtfs_realtime_version", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "incrementality", kind: "enum", opt: true, T: () => ["transit_realtime.FeedHeader.Incrementality", FeedHeader_Incrementality] },
            { no: 3, name: "timestamp", kind: "scalar", opt: true, T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 4, name: "feed_version", kind: "scalar", opt: true, T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<FeedHeader>): FeedHeader {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.gtfsRealtimeVersion = "";
        if (value !== undefined)
            reflectionMergePartial<FeedHeader>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: FeedHeader): FeedHeader {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* required string gtfs_realtime_version */ 1:
                    message.gtfsRealtimeVersion = reader.string();
                    break;
                case /* optional transit_realtime.FeedHeader.Incrementality incrementality = 2 [default = FULL_DATASET] */ 2:
                    message.incrementality = reader.int32();
                    break;
                case /* optional uint64 timestamp */ 3:
                    message.timestamp = reader.uint64().toBigInt();
                    break;
                case /* optional string feed_version */ 4:
                    message.feedVersion = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: FeedHeader, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* required string gtfs_realtime_version = 1; */
        if (message.gtfsRealtimeVersion !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.gtfsRealtimeVersion);
        /* optional transit_realtime.FeedHeader.Incrementality incrementality = 2 [default = FULL_DATASET]; */
        if (message.incrementality !== undefined)
            writer.tag(2, WireType.Varint).int32(message.incrementality);
        /* optional uint64 timestamp = 3; */
        if (message.timestamp !== undefined)
            writer.tag(3, WireType.Varint).uint64(message.timestamp);
        /* optional string feed_version = 4; */
        if (message.feedVersion !== undefined)
            writer.tag(4, WireType.LengthDelimited).string(message.feedVersion);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message transit_realtime.FeedHeader
 */
export const FeedHeader = new FeedHeader$Type();
// ...
export const FeedEntity = new FeedEntity$Type();
// ...
export const TripUpdate = new TripUpdate$Type();
// ...
export const VehiclePosition = new VehiclePosition$Type();
// ...
export const Alert = new Alert$Type();
// ...
export const Shape = new Shape$Type();
// ...
export const Stop = new Stop$Type();
// ...
export const TripModifications = new TripModifications$Type();